scripts/requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
numpy>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
mysql-connector-python>=8.3.0
sec-edgar-api>=1.0.4
tqdm>=4.66.1
